✨ feat(user): added voters

+++ src/Domain/User/Action/DeleteUserAction.php
@@ -8,15 +8,16 @@ use App\Application\Common\Action\BaseAction;
 use App\Application\Common\Enum\HttpMethodEnum;
 use App\Application\Common\Exception\EntityNotFoundHttpException;
 use App\Domain\User\Repository\UserRepository;
+use App\Domain\User\Security\UserVoter;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 use Symfony\Component\Routing\Attribute\Route;
 
 #[AsController]
-readonly class DeleteUserAction extends BaseAction
+class DeleteUserAction extends BaseAction
 {
     public function __construct(
-        private UserRepository $userRepository,
+        private readonly UserRepository $userRepository,
     ) {
     }
 
@@ -28,6 +29,8 @@ readonly class DeleteUserAction extends BaseAction
         if (null === $user) {
             throw new EntityNotFoundHttpException($id);
         }
+
+        $this->denyAccessUnlessGranted(UserVoter::DELETE, $user);
 
         $this->userRepository->delete($user);
